add buy car and sell car function

+++ car-inventory-service/src/main/java/com/car/microservices/carinventoryservice/repository/CarInventoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.car.microservices.carinventoryservice.dto.Car;
 
-public interface CarInventoryRepository extends JpaRepository<Car, Long>{
+public interface CarInventoryRepository extends JpaRepository<Car, Integer>{
 
 	List<Car> findByStatusIgnoreCase(String status);
 }
